services/agent-service/requirements.txt
# FastAPI
fastapi==0.115.12
uvicorn[standard]==0.34.0

# LangChain ecosystem
langchain==0.3.21
langchain-openai>=0.3.0,<0.4.0
langchain-anthropic>=0.3.3,<0.4.0

openai>=1.30.0,<2.0.0
anthropic>=0.41.0,<1.0.0
google-generativeai>=0.5.0,<1.0.0

# Agent orchestration
autogen>=0.11.0,<0.12.0

# Data processing
pandas>=2.2.1,<2.3.0
numpy>=1.26.4,<1.27.0
beautifulsoup4>=4.12.3,<5.0.0
lxml>=5.1.0,<6.0.0
cssselect>=1.2.0,<2.0.0

# Async
aiohttp>=3.9.3,<4.0.0
aiokafka>=0.12.0,<0.13.0

# Database
sqlalchemy[asyncio]>=2.0.27,<3.0.0
asyncpg>=0.29.0,<0.30.0
redis>=5.0.1,<6.0.0

# Utilities
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-dotenv>=1.0.1,<2.0.0
structlog>=24.1.0,<25.0.0
tenacity>=8.2.3,<9.0.0
ulid-py>=1.1.0,<2.0.0
orjson>=3.9.14,<4.0.0

# Monitoring
prometheus-client>=0.19.0,<1.0.0